requirements.txt
aiogram
asyncio
psycopg2-binary
python-dotenv
watchdog
psycopg2
PyYAML
